% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_common_recodes.R
\docType{data}
\name{five_r}
\alias{five_r}
\title{Common recoding themes}
\format{
An object of class \code{numeric} of length 5.
}
\usage{
five_r
}
\description{
These are common recoding themes used in the electoralContestation package
}
\keyword{datasets}
