% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorMarginal.R
\name{posterior_pme}
\alias{posterior_pme}
\title{#' @title posterior_pme
Generate Predictive Marginal Effect}
\usage{
posterior_pme(
  model = burn_flag,
  xvar = "prepost",
  mvar = "presvote_trump_2020",
  mrange = c(0, 1),
  xrange = c("pre", "post")
)
}
\arguments{
\item{model}{{brms model}}

\item{xvar}{{independent variable}}

\item{mvar}{{moderator variable}}

\item{mrange}{{values of mvar}}

\item{xrange}{{values of xvar}}
}
\value{
R data frame with posterior predictions for each category, across levels of xvar, all other variables held at mean.
}
\description{
Use category_groups to group categories together. The function asks the user to choose how to group
the categories together, in order to aid interpretation with many category models (e.g. ordinal or multinomial models).
The function returns the average marginal effect for each category, across levels of xvar, all other variables observed.
}
