% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMeta.R
\name{print_metadata}
\alias{print_metadata}
\title{Create a zero one normalization}
\usage{
print_metadata(data)
}
\arguments{
\item{numeric}{value}
}
\value{
A standardized 0-1 variable
}
\description{
Create a zero one normalization
}
\examples{
#rnorm(10) \%>\% zero.one()
}
