% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recodeColumn}
\alias{recodeColumn}
\title{recodeColumn}
\usage{
recodeColumn(data = df, column, recode_rules, colname)
}
\arguments{
\item{data}{data frame}

\item{recode_rules}{specify the recode rules,  c("1" = 4, "2" = 3, "3" = 2, "4" = 1)}

\item{colname}{specify the new column name}

\item{id}{specify a recode column}
}
\value{
The data frame with recoded column
#'
}
\description{
Uniform recode function for survey data

#'
}
