% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_recodeIterator.R
\name{recodeList}
\alias{recodeList}
\title{Apply multiple recoding operations to a dataframe}
\usage{
recodeList(data, recode_list)
}
\arguments{
\item{data}{A dataframe.}

\item{recode_list}{A list of recoding operations. Each operation is a list containing the column, recode_rules, and new_column.}
}
\value{
A dataframe with the recoded columns.
}
\description{
Apply multiple recoding operations to a dataframe
}
